wip: crosshairs appears
